collectionsPro/List/src/test/java/list/SimpleQueueTest.java
package list;

import list.SimpleQueue;
import org.junit.Before;
import org.junit.Test;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.*;

public class SimpleQueueTest {
    private SimpleQueue<Integer> queue;

    @Before
    public void setQueue() {
        queue = new SimpleQueue<>();
        queue.push(1);
        queue.push(2);
        queue.push(3);
    }
    @Test
    public void sequentialOfPushAndPollMethodsInvocations() {
        assertThat(queue.poll(), is(1));
        assertThat(queue.poll(), is(2));
        assertThat(queue.poll(), is(3));
        assertEquals(queue.poll(), null);
    }
}